Moved method array joinStrings to ArrayUtils class from CommonUtils class and related changes code and tests
